requirements.txt
-i https://pypi.org/simple
django-filter==2.1.0
django==2.1.8
djangorestframework==3.9.4
isort==4.3.18
pymysql==0.9.3
pytz==2019.1
pip==19.0.3
